backend/requirements.txt
fastapi
uvicorn
tensorflow
pandas
numpy
scikit-learn
joblibfastapi
uvicorn
tensorflow
pandas
numpy
scikit-learn
joblib
supabase
python-dotenv
